minimalistic 2D platformer in Unity
You'll need to create a player character with adjustable speed and jump height, platforms to move and jump across, as well as some coins to collect. If the player collects all the coins, they win the game and should see the ending screen.

If the player falls and goes below the screen, they should lose a life and respawn at the beginning (but they should not lose any coins). If the player runs out of lives, they should be sent to a Game Over screen. The Game Over screen should allow the player to return to the title screen.

You need to have at least 4 scenes: Title, Gameplay, Game Over, and Ending/Victory.